fix: 副作用が出るためmixinのグローバル化を廃止
PC判定もmixinに移動
